Fixes and code enhancements
The region property of the leagueLarge schema was appearing in both
/getLive and /getLeagues endpoints when it was only required in the latter
endpoint.

Fixed issue in /getLive endpoint where it was referring to the leagueSmall
endpoint when it required the larger version. This was due to the league
property being referenced in the event schema used by both /getSchedule
and /getLive.

Added missing slug property to the team schema and id property to the
event schema. Both were intended only for the /getLive endpoint.

Since every endpoint that uses the match schema also contains the team
schema within it, the code has been changed to reflect this and reduce
redundancy.

Improved the descriptionfor the record schema

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>