Fix integer types
Initially some schemas had been incorrectly been labelled as numbers.
The number data type is used for floating point values. The correct data
type for these schemas was integer. This commit fixes that issue.

The size of the integer is only specified if it is int64. This is because
int32 is considered the default size in most implementations.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>